Metagraph constructs ZF set
Responsibility for converting a path in the metagraph (and its associated edge data) to a zero forcing set has been moved to the metagraph class.  This makes sense because how this is done depends on the type of forcing in question, and therefore depends on the particular subclass of the metagraph class that is used.