Added build-and-test workflow
Made a workflow that builds the code and tests it on
pull requests and pushes to `master`. Might have to remove the rule for
pushes to `master` but we'll cross that bridge when we get to it. Added the `cleanup` branch for now as well, but this should be replaced with some kind of `development` branch after the cleanup is performed.