Revert "Added barebones implementation of RadixTree"
This reverts commit eb3711f9b909bbfb1226b9ae6e5f562be5955e90.

Reverting because this can probably be rethought in the future, but I think I made
some incorrect assumptions with the `RadixTree` that I really don't
remember precisely.

I believe the code was working, but the tests didn't pass -- I tried
confirming that but the code didn't even compile. Either way, it's fine
to shelf for now.